- Added 3 new effects:     - Cancel Normal Effect     - Consume Item     - Damage Item - Added 9 new effect causes:     - Item Right Click     - Item Right Click Block     - Item Right Click Entity     - Item Using     - Item Left Click     - Item Digging     - Item Break Block     - Item Attack     - Item Kill - Improved item displays - Fixed crash with club

+++ src/main/java/com/tmtravlr/additions/util/client/CommonGuiUtils.java
@@ -2,7 +2,6 @@ package com.tmtravlr.additions.util.client;
 
 import java.awt.Desktop;
 import java.io.File;
-import java.net.URI;
 
 import javax.swing.JFileChooser;
 import javax.swing.UIManager;
@@ -43,7 +42,7 @@ public class CommonGuiUtils {
 	}
 	
 	public static void drawStringWithDots(FontRenderer fontRenderer, String text, int width, int x, int y, int color) {
-		fontRenderer.drawStringWithShadow(trimWithDots(fontRenderer, text, width), (float)x, (float)y, color);
+		fontRenderer.drawStringWithShadow(trimWithDots(fontRenderer, text, width), x, y, color);
 	}
     
     public static void openFolder(File folder) {
